Create project 2
This code prompts the user to input different supplier codes and the quantity of products each supplier brought today. The user can terminate the program by entering -1. Then, it calculates the total quantity of products based on the input and determines the supplier who brought the most products. The program tracks the product quantities associated with supplier codes and identifies the supplier with the highest quantity of products based on these codes.